requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
phonenumbers==8.13.24
pycountry==23.12.11
python-dotenv==1.0.0
google-auth==2.23.4
google-cloud-bigquery==3.14.1
google-api-python-client==2.108.0
google-cloud-secret-manager==2.18.1
openpyxl==3.1.2
xlrd==2.0.1
chardet==5.2.0
pyarrow==15.0.0
httpx==0.27.0
